Refactor UI utility functions into a separate module
This commit completes the refactoring of `gui/main.py` by extracting the `open_file_dialog` function into a dedicated UI utilities module.

The following changes are included:

- **New UI Utilities Module:**
  - A new file `gui/utils.py` has been created.

- **Move Function:**
  - The `open_file_dialog` function has been moved from `gui/main.py` to `gui/utils.py`.

- **Update Imports:**
  - `gui/main.py` has been updated to import and expose the function from its new location.

This final refactoring makes the responsibilities of `gui/main.py` cleaner, focusing it solely on high-level control and leaving utility functions in their own dedicated modules. This marks the completion of a comprehensive effort to improve the codebase's structure, documentation, and test coverage.
